requirements.txt
aiodiskqueue==0.1.2
Jinja2==3.1.6
mysql-connector-python==9.2.0
oracledb==2.5.1
plyer==2.1.0
psycopg2==2.9.9
psycopg2_binary==2.9.9
pymongo==4.11.2
pyodbc==5.2.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
Requests==2.32.3
sshtunnel==0.4.0
